+++ packages/store/internals/src/angular.ts
@@ -1,12 +1,16 @@
 import { getPlatform, COMPILER_OPTIONS, CompilerOptions, PlatformRef } from '@angular/core';
 import { memoize } from './memoize';
 
-// TODO: tree-shake this away in production since we're using it only to check whether
-// NGXS is running in test mode!
-function _isAngularInTestMode() {
+/**
+ * @description Will be provided through Terser global definitions by Angular CLI
+ * during the production build. This is how Angular does tree-shaking internally.
+ */
+declare const ngDevMode: boolean;
+
+function _isAngularInTestMode(): boolean {
   const platformRef: PlatformRef | null = getPlatform();
   if (!platformRef) return false;
-  const compilerOptions = platformRef.injector.get<any>(COMPILER_OPTIONS, null);
+  const compilerOptions = platformRef.injector.get(COMPILER_OPTIONS, null);
   if (!compilerOptions) return false;
   const isInTestMode = compilerOptions.some((item: CompilerOptions) => {
     const providers = (item && item.providers) || [];
@@ -20,4 +24,7 @@ function _isAngularInTestMode() {
   return isInTestMode;
 }
 
-export const isAngularInTestMode = memoize(_isAngularInTestMode);
+export const isAngularInTestMode =
+  // Caretaker note: we have still left the `typeof` condition in order to avoid
+  // creating a breaking change for projects that still use the View Engine.
+  typeof ngDevMode === 'undefined' || ngDevMode ? memoize(_isAngularInTestMode) : () => false;