+++ packages/store/src/decorators/select/symbols.ts
@@ -1,20 +1,19 @@
 import { Observable } from 'rxjs';
 
-import { CONFIG_MESSAGES, VALIDATION_CODE } from '../../configs/messages.config';
 import { propGetter } from '../../internal/internals';
 import { SelectFactory } from './select-factory';
 import { StateToken } from '../../state-token/state-token';
 import { ExtractTokenType } from '../../state-token/symbols';
+import { throwSelectFactoryNotConnectedError } from '../../configs/messages.config';
 
 const DOLLAR_CHAR_CODE = 36;
 
 export function createSelectObservable<T = any>(selector: any): Observable<T> {
   // We're not adding `ngDevMode` guard since we have still to throw this error until we fix SSR issue.
   if (!SelectFactory.store) {
-    throw new Error(CONFIG_MESSAGES[VALIDATION_CODE.SELECT_FACTORY_NOT_CONNECTED]());
+    throwSelectFactoryNotConnectedError();
   }
-
-  return SelectFactory.store.select(selector);
+  return SelectFactory.store!.select(selector);
 }
 
 export function createSelectorFn(name: string, rawSelector?: any, paths: string[] = []): any {